Decorators API Refactored
