+++ tensorlayer/decorators/__init__.py
@@ -1,0 +1,16 @@
+"""
+TensorLayer provides rich layer implementations trailed for
+various benchmarks and domain-specific problems. In addition, we also
+support transparent access to native TensorFlow parameters.
+For example, we provide not only layers for local response normalization, but also
+layers that allow user to apply ``tf.nn.lrn`` on ``network.outputs``.
+More functions can be found in `TensorFlow API <https://www.tensorflow.org/versions/master/api_docs/index.html>`__.
+"""
+
+from .deprecation import deprecated_alias
+from .private_method import private_method
+
+__all__ = [
+    'deprecated_alias',
+    'private_method'
+]
