+++ tensorlayer/decorators/deprecation.py
