+++ tensorlayer/layers/object_detection.py
@@ -3,7 +3,7 @@
 from tensorlayer import tl_logging as logging
 from tensorlayer.layers.core import *
 
-from tensorlayer.deprecation import deprecated_alias
+from tensorlayer.decorators import deprecated_alias
 
 __all__ = [
     'ROIPoolingLayer',
