+++ tensorlayer/layers/recurrent.py
@@ -8,7 +8,7 @@ from tensorflow.python.util.tf_inspect import getfullargspec
 from tensorlayer import tl_logging as logging
 from tensorlayer.layers.core import *
 
-from tensorlayer.deprecation import deprecated_alias
+from tensorlayer.decorators import deprecated_alias
 
 __all__ = [
     'RNNLayer',
