+++ tensorlayer/layers/spatial_transformer.py
@@ -7,7 +7,7 @@ from six.moves import xrange
 from tensorlayer import tl_logging as logging
 from tensorlayer.layers.core import *
 
-from tensorlayer.deprecation import deprecated_alias
+from tensorlayer.decorators import deprecated_alias
 
 __all__ = [
     'transformer',
