+++ tensorlayer/layers/stack.py
@@ -5,7 +5,7 @@ import tensorflow as tf
 from tensorlayer import tl_logging as logging
 from tensorlayer.layers.core import *
 
-from tensorlayer.deprecation import deprecated_alias
+from tensorlayer.decorators import deprecated_alias
 
 __all__ = [
     'StackLayer',
